tests/wpt/tests/html/semantics/embedded-content/the-iframe-element/iframe_sandbox_navigate_descendants.html

<!doctype html>
<meta charset="utf-8">
<title>Check that sandboxed iframe can navigate their descendants</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  var t = async_test();
  onmessage = t.step_func_done(function(e) {
    assert_equals(e.data, "can navigate", "Should have the right message");
  });
</script>
<iframe sandbox="allow-scripts" src="support/iframe-trying-to-navigate-its-child.html"></iframe>
